fix: menu 브랜치 복구
전체 복구

+++ IrumOrder_BackEnd/src/main/java/Irumping/IrumOrder/controller/MenuRequest.java
@@ -4,14 +4,20 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-
+/**
+ * 클래스 설명: 메뉴 생성 요청 정보를 담는 클래스
+ * 사용자가 입력한 메뉴 이름, 가격, 카테고리 아이디를 담는다.
+ *
+ * 작성자: 주영은
+ * 마지막 수정일: 2024-12-04
+ */
 @Getter
 @Setter
 @RequiredArgsConstructor
 public class MenuRequest {
 
     private String name;
-    private Long price;
-    private Long categoryId;
+    private Integer price;
+    private Integer categoryId;
 
 }
